Created util module
